Revert "test continuous memory"
This reverts commit bbb20e338341f35bb46f3c1c92b590a7cb974905.

+++ src/Grid.h
@@ -27,7 +27,7 @@ public:
 	vector<float>* getDistances(int i);
 
 private:
-	vector<vector<float> > grid;
+	vector<vector<int> > grid;
 	vector<vector<int> > speedOctopus;
 
 	vector<vector<int>*> neighbors;
